+++ Assets/Scripts/ManagersAndFactories/TankPartSchematicFactory.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 
 public class TankParSchematictFactory
 {
-    public static WheelPartSchematic CreateWheelPartSchematic(string name, float energyInc, float energyDec, float weight, KeyCode leftForwardKey=KeyCode.None, KeyCode leftBackKey=KeyCode.None, 
+    public static WheelPartSchematic CreateWheelPartSchematic(string name, float energyInc, float energyDec, int weight, KeyCode leftForwardKey=KeyCode.None, KeyCode leftBackKey=KeyCode.None, 
         KeyCode rightForwardKey = KeyCode.None, KeyCode rightBackKey = KeyCode.None) {
 
         if (leftForwardKey == KeyCode.None) {
@@ -28,11 +28,11 @@ public class TankParSchematictFactory
         return new WheelPartSchematic(name, energyInc, energyDec, weight, leftForwardKey, leftBackKey, rightForwardKey, rightBackKey);
     }
 
-    public static HullPartSchematic CreateHullPartSchematic(string name, int armour, Vector2 size, float energyPower, float weight) {
+    public static HullPartSchematic CreateHullPartSchematic(string name, int armour, Vector2 size, int energyPower, int weight) {
         return new HullPartSchematic(name, armour, size, energyPower, weight);
     }
 
-    public static TurretPartSchematic CreateTurretPartSchematic(string name, int armour, float rotPerTimeStep, float weight, Vector2[] weaponDirs, float[] weightRestrict, 
+    public static TurretPartSchematic CreateTurretPartSchematic(string name, int armour, float rotPerTimeStep, int weight, Vector2[] weaponDirs, int[] weightRestrict, 
         KeyCode leftTurnKey = KeyCode.None, KeyCode rightTurnKey = KeyCode.None, KeyCode[] shootKeys = null) {
 
         if (shootKeys == null) {
@@ -52,7 +52,7 @@ public class TankParSchematictFactory
         return new TurretPartSchematic(name, armour, rotPerTimeStep, weight, weaponDirs, weightRestrict, leftTurnKey, rightTurnKey, shootKeys);
     }
 
-    public static WeaponPartSchematic CreateWeaponPartSchematic(string name, float shootingForce, float reloadTime, float range, float weight, Bullet.BulletTypes bulletType, int damage) 
+    public static WeaponPartSchematic CreateWeaponPartSchematic(string name, float shootingForce, float reloadTime, float range, int weight, Bullet.BulletTypes bulletType, int damage) 
     {
         return new WeaponPartSchematic(name, shootingForce, reloadTime, range, weight, bulletType, damage);
     }
